requirements.txt
python-docx==0.8.11 
python-dotenv
requests 